% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{estimate.ldl}
\alias{estimate.ldl}
\title{Predict LDL levels}
\usage{
estimate.ldl(data)
}
\arguments{
\item{data}{Data.frame or matrix where the first column is what you want to predict, and remaining columns are what you want to consider as possibly covariates}
}
\description{
This function uses LASSO to select predictors in order to run linear regression to predict LDL for individuals. Outputs a vector of predicted values of LDL based on input
}
\examples{
cat_function()
}
\keyword{preds}
